requirements.txt
bleach==6.3.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
click==8.3.1
Flask==3.1.2
Flask-CKEditor==1.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
webencodings==0.5.1
Werkzeug==3.1.4
WTForms==3.2.1
